+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * added query stats
+* added lambda on exception in retryTransaction
 
 ## 1.10.0
 * changed level of update token log record from info to debug